Clean up error handling logic
The previous error handling logic was itself throwing an error when the parser
returned a `nil` result without a failure reason, which appaers to happen when
the input is missing a newline at the end of the file. This may actually be an
error in the treetop parser itself, but the changes in this commit update our
error logic so that it at least provides a semi-useful error message when this
`nil` result scenario occurs.